REST-client: add NotFoundError
If we query a specific object (like a resource-definition), we might get
an error for various reasons. It could be something like an internal
server error, but it also could be a 404 messaging that the specific
resource we queried does not exist (e.g.,
/resource-definitions/doesnotexist).

Return a specific error type in this case. Users interested in details
have to check for client.NotFoundError.