tests: clean up in Snapshot tests
In the tear down phase we can not "sp d", if there are still resources
using it. In this case the Snapshot tests did not clean up their test
resource.

We could do the resource clean up in the tear down, but it feels cleaner
that tests clean up after themselves.