REST-client: check that data fields are not nil
So far we faild parsing the json data from LINSTOR if the json did
contain a "type", but no valid "data". Check for that specifically.

This means that in these situations (where no data is set), one gets a
result with null values.

We might decide that missing "data" alrady is an error.

LINBIT/linstor-csi#28