generalize logging interface
Introduce Logger and LeveledLogger, where Logger is implemented by a
rudimentary logging library (i.e. Go's "log" package) and a
LeveledLogger is implemented by a more sophisticated library like
logrus.

This is the same principle as was recently introduced in
go-retryablehttp by hasicorp, so credits for the idea go to the author
of the pull request:

https://github.com/hashicorp/go-retryablehttp/pull/75