client: add Is() to check errors using mask values
Note that this is only really useful for the Fail* masks, the others
would be more complicated. For example, the mask for "error" is
0xC000000000000000, and for "info" it is 0x4000000000000000. In binary,
that would be:

error:	1100 0000 0000 0000 ...
info:	0100 0000 0000 0000 ...

So using the kind-of naive `(retcode & mask) == mask` that we do would
say Is(info) == true for an rc-code that is really an error.

To fix that, we would need to actually split up the rc code like it was
meant to -- into message type, operation type, object type, and the
error code.