monitor: add resource lost
This adds a new monitor package and a (New)LostResourceUser. It returns
the names of DRBD resources that have lost their user and may be
promoted to Primary on a Go channel. This "observer" keeps track of
events and gives resources a configurable grace period before reporting
them as lost.