Issues: #10
Fix the bug that can't get storage_pool by a volume from the GetResourceView.

When I use Resources.GetResourceView() to get the resource list with
volume， but can't get the volume.StoragePool, the value is always "".
And then I find the api response is storage_pool_name from /v1/view/resources.
but in our code, it use storage_pool to encode storage_pool_name, so it's
not valid.

Then I fix the bug by change the storage_pool to storage_pool_name.