fix parsing of volume responses
Parsing of volumes relies on the existing (generated) enum names to align
with the response of the Linstor API. The generated enum values did not
completely align with the Linstor API, so normal responses ended up not
parsing.

The fix is to change the enum values in the linstor-common project, as they
should always align with the Linstor API. A rerun of go generate fixes the
issue. To ensure parsing of those resources does not break again, some new
tests were added.

This bug was introduced in 110e5f8e9cf while switching to the new style
enum generation. Before that commit, the previous values were the once
returned by Linstor.