client: support for multiple controller URLs
This works the same as linstor-api-py's MultiLinstor. When Connect
is called, golinstor tries to connect to each potential controller URL
and uses the first working URL that is found.

It also implements a cache mechanism for the base URL so that the
correct controller does not have to be determined every time. If a
request fails (e.g. because the controller moves to another node), the
controller list is scanned again.