client: make volume number optional
When the volume number is omitted from a creation request, LINSTOR just
picks the next available volume number. This is fine, but this does not
cover the case when we actually want to specify volume number zero.

So, make the volume number a pointer so we can actually specify an
intended empty value.