http: optionally set user agent
For better traceability, it is a good idea to include a user agent in requests.
LINSTOR does report the user agent of a request that caused an error.

Since golinstor itself is only a library, the users of the library should be
able to set the User Agent header to their own values.