remove unused client and test code
The tests did not compile starting with 3fae30daadef, maybe even
earlier. Nobody noticed, because nobody is running these behaviour
tests.

The tests are all just basic behaviour of LINSTOR and not specific
to golinstor itself, so removing them should be fine.

Secondly, the "alternative" client in linstor.go is similarly unused.
Since that client expects to actually call the "linstor" python client,
we should be free to remove it as well: nobody is using it.