client: export BaseURL via getter
It is now possible to pass a list of controllers to a Client (via
LS_CONTROLLERS). This means that it may not always be obvious from
"the outside" which LINSTOR controller we are actually talking to.

For use cases that want to know this, we could just export the baseURL
member of Client. But I don't think we want to enable users to change
the base URL of an existing Client, so let's just add a getter function
to make it read-only.