client: add option to directly set rate.Limiter
In some contexts it may be necessary to create new clients while keeping
a global rate limit on the API. Such is the case for the k8s operator.

In order to have a single rate.Limiter instance, we need to be able
to set the Limiter directly instead of manually constructing it via
Limit().