add the option to do client side caching
We have some customers that run pretty big setups with lots of nodes and
resources. LINSTOR will always try to show up-to-date information
(as it should), which makes many simple "GET" requests quite expensive on
these clusters.

This is particularly true for things like LINSTOR CSI and the LINSTOR Operator,
which can potentially create many similar requests, for slightly different data
in parallel. LINSTOR already provides "cluster view" endpoints for those, but
that needs client side caching in both projects.

For LINSTOR CSI we already implemeneted the relevant caching strategies, but
we now have a case where this would also be useful to have in the Operator.
To that end, we "promote" the client side caching from LINSTOR CSI to
golinstor.

The cache implementation hooks into the Node and Resource providers and
caches everything base on the a single "view" request. It will try to
serve all further requests from the cached view, until either the timeout
expired, or the client made a modification.