cache: fix filtering by properties
While initially assumed to be unimportant, turns out that this is actually
used in some code paths. So we implement the necessary filtering by properties.

There are two forms to consider:
* filter by key and values, by using key=value as filter
* filter by key, by using just "key" as filter

While implementing this, it also turns out that the way nodes and storage pools
are filtered was inconsistent with LINSTOR behaviour: LINSTOR will return a
result if any item matches, not just if all items match. This has been fixed now,
too.