cache: invalidate cache only after operation
The cache is most useful in multi-threaded scenarios. If the cache is
invalidated before the modification is applied, there is a chance it
gets populated by another go routine before the change can be observed.

To fix this, defer calls to .Invalidate(). This ensures any call that is
part of the same goroutine will see (at least) the specific modification
made by the call.