fix checking for specific errors
The way we checked for specific error responses by LINSTOR was broken:
because some errors have overlapping bitmaps, the existing check would
report that a specific error occured, even if actually a different error
occured.

For example, checking for FailExistsCryptPassphrase would succeed
for all other error code > 512 and < 1024, as well.

Fix this by using the correct bit mask to extract the errors.